+++ src/core/lispStream.cc
@@ -5071,12 +5071,12 @@ void writeln_bf_stream(const std::string& fmt, T_sp strm)
 
 CL_LAMBDA(oject stream);
 CL_DECLARE();
-CL_DOCSTRING("Write the address of an object to the stream.");
+CL_DOCSTRING("Write the address of an object to the stream designator.");
 DOCGROUP(clasp);
 CL_DEFUN void core__write_addr(T_sp x, T_sp strm) {
   stringstream ss;
   ss << (void *)x.raw_();
-  writestr_stream(ss.str().c_str(), strm);
+  writestr_stream(ss.str().c_str(), coerce::outputStreamDesignator(strm));
 }
 
 static cl_index